Move the StatefullRadioGroup to stateful/RadioGroup
